Vector3 -> Vector3 and Point3
The class Vector3 which describes also a point leads to a lot of
troubles.
Mostly, we need to check is something vector or a point. Sometimes, the
method should not work with a point (e.t. point+point). But we cannot
let user know this.
So I decided to rewrite Point and Vector as separate classes.
Also, there is no conversions between point and vector other than '+'
and '-' operators to not get confused.